feat: auth middleware
Next.js middleware as of now requires it to be run on the Edge runtime. This runtime has limitations, including not being able to access crypto libraries needed for `bcrypt`, nor being able to run database operations (prisma). This kinda leaves us without a lot of options, if we want our middleware to perform auth checks. Of these sub-optimal options, the best of which, IMO, is to run "middleware" directly from the routes themselves rather than implicitly using Next.js.

So create our own auth middleware library which will be used within our protected routes. This middleware confirms the auth cookie exists and is populated with a user's `userId`. It returns a `Session` which contains only the `User`, or throws `UnauthorizedError` if this is not the case.