feat: cuid as primary key (#27)
To avoid leaking information to the client about our data, switch to using cuid2 as the primary key for the database tables. This allows us to remove the `userId` field from User and use the `id` field instead, similar steps done for Widget. (We also must update to Prisma 6.x to use cuid2.)

There are lots of opinions on how to structure primary keys and how to expose public IDs for clients. In the end, this seemed like the simplest approach, in that we do not need to create both a sequential id and a cuid for a public ID, but instead use the cuid as the unique ID. The downsides of using cuid as a primary key have mostly gone away as the hardware ability has improved.

An aside: we replaced the internal use of `nanoid` by our logging infrastructure with a direct dependency on `cuid2`, which is why we have it in our npm dependencies now.